Remove toggle cost from Hail Mary protocol
It is a remnant from when fuel costs were introduced (in 2023).

This commit instead adds it to the trigger cost that was introduced
more recently (2024), because no other controls have a toggle cost and
are instead removing fuel only when triggered.

Therefore this commit brings the Hail Mary control in line with all the
other controls.