fix console part eating cartridges when swapping
When a cartridge already is in the console port, then putting in another 
cartridge will eject the previous one.
Unfortunately trying to eject the new cartridge now (i.e. without swapping), 
then this will not work and instead the cartridge will be gone for good.

This commit moves the call that clears the hasCartridge state to a different 
location in the, so that cartridges only then get cleared when there is no 
cartridge in the port anymore (incl. when swapping in a different cartridge).