fix console generator breaking in creative mode
When trying to switch to another console while in creative mode, 
left-clicking to browse left will break the console generator instead.
This commit makes the left/right click behavior of the console generator 
be consistent between creative mode and survival mode, so that players 
don't accidentally break the console generator when they just wanted 
to switch the console.