clean up crash effects
Its fire griefing check has become obsolete.
It also has sounds that are not played by default and don't 
make sense anymore to begin with.
Instead this commit moves the playSoundAtConsole call into the 
crash effects method.

This way the method can be used to expand on the crash effects in the future.